Refactor desktop-app and adjust app size
Refactored 'desktop-app.go' under 'internal/desktop-app' directory for code clarity and consistency. The package and function name have been renamed from 'desktop_app' and 'TestDesktop()' to 'desktopApp' and 'RunDesktop()', respectively, aligning better with Go naming conventions. Additionally, the window size for the application has been expanded from 250 to 285 for better UI experience.
